+++ CHANGELOG.md
@@ -8,8 +8,11 @@ ___Note:__ Yet to be released changes appear here._
 
 ### General
 
-* `FIX`: show default context menu for text inputs in diagram ([#5803](https://github.com/camunda/camunda-modeler/issues/5803))
 * `FEAT`: show warning on selected version and connection version mismatch ([#5652](https://github.com/camunda/camunda-modeler/issues/5652))
+* `FIX`: show default context menu for text inputs in diagram ([#5803](https://github.com/camunda/camunda-modeler/issues/5803))
+* `FIX`: correct bug opening DMN decision table ([#5880](https://github.com/camunda/camunda-modeler/issues/5880), [#5898](https://github.com/camunda/camunda-modeler/pull/5898))
+* `CHORE`: track variable outline panel open/close events ([#5878](https://github.com/camunda/camunda-modeler/pull/5878))
+* `DEPS`: update to `@camunda/linting@3.49.0`
 
 ## 5.46.1